requirements.txt
autopep8==1.5.7
backcall==0.2.0
click==8.0.1
decorator==5.0.9
environs==9.3.2
Flask==2.0.1
Flask-Cors==3.0.10
gunicorn==20.1.0
ipdb==0.13.9
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.12.1
matplotlib-inline==0.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
python-dotenv==0.17.1
six==1.16.0
toml==0.10.2
traitlets==5.0.5
wcwidth==0.2.5
Werkzeug==2.0.1
